+++ packages/plugin-packages/mashroom-background-jobs/src/js/context/global_context.ts
@@ -1,0 +1,8 @@
+
+import MashroomBackgroundJobPluginRegistry from '../plugins/MashroomBackgroundJobPluginRegistry';
+
+const pluginRegistry = new MashroomBackgroundJobPluginRegistry();
+
+export default {
+    pluginRegistry,
+};