+++ packages/plugin-packages/mashroom-background-jobs/src/js/services/mashroom-bootstrap-services.ts
@@ -1,0 +1,14 @@
+
+import MashroomBackgroundJobService from './MashroomBackgroundJobService';
+
+import type {MashroomServicesPluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
+
+const bootstrap: MashroomServicesPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
+    const service = new MashroomBackgroundJobService(pluginContextHolder);
+
+    return {
+        service,
+    };
+};
+
+export default bootstrap;