+++ packages/plugin-packages/mashroom-background-jobs/src/js/webapp/routes/index.ts
@@ -1,0 +1,6 @@
+
+import type {Request, Response} from 'express';
+
+export default (req: Request, res: Response) => {
+    res.redirect(`${req.baseUrl}/admin`);
+};