+++ packages/plugin-packages/mashroom-portal-remote-app-registry/src/js/services/MashroomPortalRemoteAppEndpointService.ts
@@ -36,7 +36,7 @@ export default class MashroomPortalRemoteAppEndpointService implements MashroomP
                 portalApps: []
             });
 
-            context.backgroundJob.runASAP();
+            context.backgroundJob.run();
         }
     }